requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
sqlalchemy==2.0.44
geoalchemy2==0.18.1
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic-settings==2.12.0
python-multipart==0.0.20
geojson==3.2.0
geopandas==1.1.1
shapely==2.1.2
fastkml==1.4.0
gpxpy==1.6.2
pandas==2.3.3
python-dotenv==1.2.1
alembic==1.17.2
pyproj==3.7.2

